requirements.txt
Django
gunicorn
djangorestframework
Pillow
# psycopg2-binary  # Se for usar PostgreSQL
